+++ urllib3/contrib/pyopenssl.py
@@ -59,7 +59,7 @@ except ImportError:  # Platform-specific: Python 3
 
 import logging
 import ssl
-import six
+from ..packages import six
 import sys
 
 from .. import util