+++ CHANGES.rst
@@ -15,6 +15,8 @@ dev (master)
 
 * Fixed compatibility for cookiejar. (Issue #1229)
 
+* pyopenssl: Use vendored version of ``six``. (Issue #1231)
+
 * ... [Short description of non-trivial change.] (Issue #)